% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchJurisdictionInfo.R
\name{fetchJurisdictionInfo}
\alias{fetchJurisdictionInfo}
\title{Fetch up to date Jurisdiction information}
\usage{
fetchJurisdictionInfo(thisTaxon = NULL, trace = FALSE)
}
\arguments{
\item{trace}{Logical. If TRUE then debug messages are printed to the console; if FALSE (default) then only sparse progress messages are output.}

\item{theseTaxa}{Character object containing the taxon name to be processed.}
}
\value{
A one row data.frame with the following columns:
\item{taxonGUID}{The ALA globally unique identifier for the taxon}
\item{taxon}{Taxonomic name whose distribution information has been extracted}
\item{isAccepted}{Logical value: Is the taxon an accepted name according to the Australian Plant Census (APC)?}
\item{isNaturalised}{Logical value: TRUE if it has a status indicating it is \emph{naturalised} in some way in \emph{any} of the jurisdictions}
\item{ACT}{Status of the taxon in the Australian Capital Territory}
\item{AR}{Status of the taxon in the Ashmore Reef territory}
\item{ChI}{Status of the taxon on Christmas Island}
\item{CoI}{Status of the taxon on the Cocos (Keeling) Islands}
\item{CSI}{Status of the taxon on the Coral Sea Islands}
\item{HI}{Status of the taxon on Heard Island}
\item{LHI}{Status of the taxon on Lord Howe Island}
\item{MDI}{Status of the taxon on McDonald Island}
\item{MI}{Status of the taxon on Macquarie Island}
\item{NI}{Status of the taxon on Norfolk Island}
\item{NSW}{Status of the taxon in New South Wales}
\item{NT}{Status of the taxon in the Northern Territory}
\item{Qld}{Status of the taxon in Queensland}
\item{SA}{Status of the taxon in South Australia}
\item{Tas}{Status of the taxon in Tasmania}
\item{Vic}{Status of the taxon in Victoria}
\item{WA}{Status of the taxon in Western Australia}
The status of a taxon may have one of the following values:
\item{native}{Occurs naturally within the jurisdiction}
\item{native and naturalised}{Occurs naturally in some part of the jurisdiction but has become naturalised in one or more other parts}
\item{naturalised}{Established self-sustaining populations occur within the jurisdiction}
\item{sparingly naturlised}{Established self-sustaining low abundance or low density populations occur within the jurisdiction}
\item{doubtfully naturalised}{Taxon has been reported as naturalised but this status is considered doubtful after expert review}
\item{formerly naturalised}{Taxon was naturalised but is now no longer considered present}
\item{presumed extinct}{Taxon did occur naturally but is now conidered extinct within the jurisdiction}
\item{NA}{There is no record of the taxon from the jurisdiction}
This function may be called independently for fun or profit, but its main role is to support the filtering of records by the function \code{\link{filterALA}}.
}
\description{
Makes calls to the Atlas of Living Australia (ALA) API to obtain information on the Australian states and territores a species has been accepted by experts to occur. These data are part of the Australian Plant Census (APC), here they are used to perform a stage of ALA occurrence record filtering in the function \code{\link{filterALAdata}}.
}
\examples{
\dontrun{
## Simple calls to fetch information:
## Nasty weed
ans <- fetchJurisdictionInfo("Digitaria sanguinalis")
## Widespread true blue native
ans <- fetchJurisdictionInfo("Dodonaea viscosa")
## Naughty native
ans <- fetchJurisdictionInfo("Eucalyptus botryoides")}
}
