% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterALA.R
\name{filterALAdata}
\alias{filterALAdata}
\title{Filter ALA records}
\usage{
filterALAdata(
  taxa = NULL,
  baseDataPath = NULL,
  recType = c("herbariumRecords", "humanObservations", "surveyRecords"),
  removeDuplicates = TRUE,
  removeMissingCoords = TRUE,
  filterByJurisdiction = TRUE,
  filterCultivated = TRUE,
  doNameCheck = TRUE,
  trace = FALSE
)
}
\arguments{
\item{taxa}{A vector of one or more \emph{ALA-accepted} species names to be processed.}

\item{baseDataPath}{The full path to the folder in which species subfolders may be found which contained output from \code{\link{fetchALAdata}}.}

\item{recType}{Which type of records will be processed? Must be one of "herbariumRecords", "humanObservations", or "surveyRecords". Abbreviations may used.}

\item{removeDuplicates}{Logical. If TRUE (default), then duplicate records are removed.}

\item{removeMissingCoords}{Logical. If TRUE (default), remove records with missing coordinates}

\item{filterByJurisdiction}{Logical. If TRUE (default), use data showing the Australian jurisdictions in which APC says the taxon is found to eliminate spurious records.}

\item{filterCultivated}{Logical. If TRUE (default), remove occurrence records identified as belonging to cultivated specimens.}

\item{doNameCheck}{Logical. Should a check of taxonomic names be performed? Default is TRUE; if FALSE, then it is assumed that names are valid and accepted as reported by \code{\link{checkTaxonName}}}

\item{trace}{Logical. If TRUE, output some hopefully useful diagnostic information to the console. Default is FALSE (i.e. whisper quiet).}
}
\value{
None
}
\description{
Applies one or more filters to previously downloaded ALA data
}
\details{
{
Data sourced from the Atlas of Living Australia (ALA) needs to be cleaned and reviewed before it is fit for use. This function allows data files output by the sister function \code{\link{fetchALAdata}} to be passed through a selection of automated preliminary filters. Users may filter:
\itemize{
\item missing or corrupted geo-coordinates
\item cultivated specimens (sadly, some herbaria include samples taken from cultivated plants and include them in normal voucher collection data! Canberra and Melbourne are serial offenders here)
\item duplicated records
\item records from jurisdictions for which APC says a taxon does \emph{not} occur (calls \code{\link{fetchJurisdictionInfo}}).}
Any combination of filters may be applied.
This function allows you to practice safe data use by applying some sensible filtering processes. However good they may be, automated filters may miss some bad records. \emph{You should \bold{always} review the output from this function before using the data.}
}
}
\examples{
\dontrun{
## Just one taxon, do all filters default setting:
filterALAdata("Acacia linifolia", "herbariumRecords", "C:/Documents/myALAstuff")

## A larger set of taxa (order is not important), just remove duplicates:
filterALAdata(c("Acacia linifolia", "Acacia decora", "Zieria smithii"),
                 "C:/Documents/myALAstuff",
                 removeMissingCoordinates = FALSE, filterByJusrisdication = FALSE,
                 filterCultivated = FALSE)}
}
