% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatTaxonName.R
\name{formatTaxonName}
\alias{formatTaxonName}
\title{Format taxon name string returned by ALA4R::speciesInfo()}
\usage{
formatTaxonName(thisNameStr)
}
\arguments{
\item{thisNameStr}{Character. String from the field taxonConcept$nameFormatted in the object returned by ALA4R::speciesInfo()}
}
\value{
A string representing the full taxon name formatted with simple mark-up (e.g. simple HTML <i></i> tags)
}
\description{
Format taxon name string returned by ALA4R::speciesInfo()
}
