% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonNameParts.R
\name{taxonNameParts}
\alias{taxonNameParts}
\title{Parse a taxonomic name into components}
\usage{
taxonNameParts(taxonName, verbose = FALSE)
}
\arguments{
\item{taxonName}{Character variable storing a name to be parsed.}

\item{verbose}{Logical. Produce diagnostic information when TRUE; default is FALSE}
}
\value{
A named character vector with elements "genus", "species", "infraRank", "infraName"
}
\description{
Parse a taxonomic name into components
}
\examples{
\dontrun{
## A genus only:
ans <- taxonNameParts("Eucalyptus")
## 'Eucalyptus sp.'
## A binomial
ans <- taxonNameParts("Acacia terminalis")
## A trinomial
ans <- taxonNameParts("Acacia terminalis subsp. terminalis")
}
}
