% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchALAdata.R
\name{fetchALAdata}
\alias{fetchALAdata}
\title{Fetch ALA records for one or more taxa}
\usage{
fetchALAdata(
  taxonList = NULL,
  baseOutputPath = defaultOutputFolder,
  theseFields = stdFields,
  doNameCheck = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{taxonList}{A character vector holding a set of taxon names to be processed.}

\item{baseOutputPath}{Character string. A path to the base folder into which output will be written. An attempt will be made to create the path to the base folder if does not already exist. A sub-folder named for each taxon in taxonList will be created and dowloaded files written into it.}

\item{theseFields}{Character vector. A set of ALA occurrence field names to be returned in the API 'GET' call. The default is a set suited to post-processing of plant occurrence records. See Details below.}

\item{doNameCheck}{Logical. Should a check of taxonomic names be performed? Default is TRUE; if FALSE, then it is assumed that names are valid and accepted as reported by \code{\link{checkTaxonName}}.}

\item{verbose}{Logical. Should additional progress messages be produced?}
}
\value{
None
}
\description{
Download records from the ALA database using API calls. For each taxon, four files are written: the raw ALA data, and a file for each of herbarium, human observation and NSW survey records.
}
\details{
{
The function \emph{fetchALAdata} handles the process of fetching occurrence data from ALA for one or more taxa. For each taxon, the function saves four csv files:
\itemize{
\item raw data
\item raw herbarium (voucher) data
\item raw human observation data
\item raw NSW Vegetation Survey data (extracted from the human observation records using a cunning bit of jiggery-pokery)}

These files can be used for whatever purpose you have mind, or they can be passed through the companion function \code{\link{filterALAdata}} to make a first pass at cleaning the data. Let me perfectly frank, \emph{ALA data is full of detritus and needs to be thoroughly cleaned before use!}

Occurrence fields: The parameter \emph{theseFields} allows you to specify the occurrence fields to be supplied in the resulting data table. The default is set to a list of fields which have proven useful in post-processing plant occurrence records. The list of available fields is long and includes many options which you may wish to use in your own spin on filtering and processing the resulting data table.

See help for \link{showOccFields} for information on available fields and their definitions.
}
}
\examples{
\dontrun{
## Download or refresh all data for a single taxon
 fetchALAdata("Acacia linifolia", baseOutpath = "/home/peterw/Restore and Renew/newData/")

## For a number of species
 fetchALAdata(c("Acacia linifolia","Wilkiea hugeliana","Banksia ericifolia","Angophora costata"),
              baseOutpath = "/home/peterw/Restore and Renew/newData/")}
}
